+++ snuba/datasets/cdc/groupassignee_entity.py
@@ -1,9 +1,7 @@
 from typing import Mapping, Sequence
 
 from snuba.datasets.entity import Entity
-from snuba.datasets.pipeline.single_query_plan_pipeline import (
-    SingleQueryPlanPipelineBuilder,
-)
+from snuba.pipeline.single_query_plan_pipeline import SingleQueryPlanPipelineBuilder
 from snuba.datasets.plans.single_storage import SingleStorageQueryPlanBuilder
 from snuba.datasets.storages import StorageKey
 from snuba.datasets.storages.factory import get_cdc_storage
