+++ snuba/datasets/entities/errors.py
@@ -5,9 +5,7 @@ from snuba.clickhouse.translators.snuba.mappers import ColumnToFunction
 from snuba.clickhouse.translators.snuba.mapping import TranslationMappers
 from snuba.datasets.entity import Entity
 from snuba.datasets.plans.single_storage import SingleStorageQueryPlanBuilder
-from snuba.datasets.pipeline.single_query_plan_pipeline import (
-    SingleQueryPlanPipelineBuilder,
-)
+from snuba.pipeline.single_query_plan_pipeline import SingleQueryPlanPipelineBuilder
 from snuba.datasets.storages import StorageKey
 from snuba.datasets.storages.errors import promoted_tag_columns
 from snuba.datasets.storages.factory import get_writable_storage
